+++ env/code/main.js
@@ -1,4 +1,6 @@
-const player = add([
+kaboom();
+
+add([
 	sprite("mark"),
 	pos(40, 20),
 	area(),