+++ examples/audio.js
@@ -1,7 +1,6 @@
 // audio playback & control
 
 kaboom({
-	global: true,
 	fullscreen: true,
 	scale: 2,
 	clearColor: [0, 0, 0, 1],