+++ examples/bench.js
@@ -1,6 +1,7 @@
 // bench marking sprite rendering performance
 
 const k = kaboom({
+	noGlobal: true,
 	fullscreen: true,
 	scale: 2,
 });