+++ examples/burp.js
@@ -1,4 +1,8 @@
-const k = kaboom();
-k.add([k.text("press 'b' to burp")]);
-k.keyPress("b", k.burp);
-k.mouseClick(k.burp);
+kaboom();
+
+add([
+	text("press 'b' to burp"),
+]);
+
+keyPress("b", burp);
+mouseClick(burp);