+++ examples/flappy.js
@@ -1,5 +1,4 @@
 kaboom({
-	global: true,
 	scale: 3,
 	fullscreen: true,
 	clearColor: [0, 0, 0, 1],