+++ examples/tiled.js
@@ -1,7 +1,6 @@
 // tiled sprite
 
 kaboom({
-	global: true,
 	fullscreen: true,
 	scale: 2,
 });