+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.8 burp() (unreleased)
 - added `burp()` for easy burping
+- kaboom functions all now global by default, toggle off by `kaboom({ noGlobal: true })`
 - added input events `touchStart(id, pos)`, `touchMove(id, pos)`, `touchEnd(id, pos)`, `mouseMove(pos)`
 - added `mouseDeltaPos()`
 - added `touchToMouse` to control if touch events should be translated to mouse events